lucas-guimaraes-borges/projeto-1/pyproject.toml
[project]
name = "projeto-1"
version = "0.1.0"
description = "Projeto Individual 1: Agente de IA Orientado a Problema"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=1.2.23",
    "langchain-openai>=1.1.12",
    "openai>=2.30.0",
    "pinecone>=8.1.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = ["pytest>=8.3"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
